refactor: extract Zai shared module to eliminate provider duplication
zai.ex and zai_coder.ex were 530-line near-copies differing only
in module name, provider ID, and base URL.

Extract shared logic to Zai.Shared, leaving thin delegation wrappers.
1059 → 556 lines (47% reduction).
